Rename {pathidx => pathpool}.rs
