+++ src/pathpool.rs
